requirements.txt
numpy
matplotlib
tkinter

Optional:
pyqt5
pyovito

